Fix how fixed-object data is imported
Alas!  No objects marked as “fixed” in section 7 of the data file by
having -1 in the third field were actually getting set as fixed in the
game because I was comparing the tuple ``(-1,)`` to ``-1`` which are
never equal.  I’m surprised as much of the game worked as it did!

Fixes #21, and allows the special case introduced in #22 to be backed
out.