Avoid using locally-defined callbacks for `yesno`.
Since most of the callbacks do not rely on the local variables at all,
there's little value in using locally defined callbacks over instance
methods.

This enables the ability to save and restore the state of the game even
in the middle of a question (including the initial "WOULD YOU LIKE
INSTRUCTIONS?")

While this is not generally possible in the normal play, this change
enables using this implementation of python-adventure in ephemeral
environments where players bring their own state.